+++ server/src/services/Store.service.ts
@@ -4,12 +4,19 @@ import { InjectRepository } from "@nestjs/typeorm";
 import { Store } from "domain/entities/Store.entity";
 import { Repository } from "typeorm";
 import { IStoreService } from "domain/interfaces/IStore";
+import { UserService } from "./User.service";
+import { User } from "domain/entities/User.entity";
+import { Category } from "domain/entities/Category.entity";
+import { CategoryService } from "./Category.service";
+import { PartialUserDto } from "domain/dto/User.dto";
 
 
 @Injectable()
 export class StoreService implements IStoreService {
   constructor(
     @InjectRepository(Store) private storeRepo: Repository<Store>,
+    private userService: UserService,
+    private categoryService: CategoryService,
   ) { }
 
   private readonly logger = new Logger(StoreService.name);
@@ -24,23 +31,32 @@ export class StoreService implements IStoreService {
   //Get a store by id
   async getOneById(id: number): Promise<Store> {
     this.logger.log(`Get a specific Store ${id}.`);
-    const store = this.storeRepo.findOneBy({ id });
-    if (!store) throw new NotFoundException(`Loja ${id} não encontrada!`);
+
+    const store = await this.storeRepo.findOneBy({ id });
+    if (!store){
+      throw new NotFoundException(`Loja ${id} não encontrada!`);
+    }
+
+    const owner = await this.verifyOwner(store.owner);
+
+    console.log(owner);
 
     return store;
   }
 
   //Create a store
-  async create(store: CreateStoreDto): Promise<Store> {
-    this.logger.log(`Creating a Store: ${store}`);
+  async create(storeDto: CreateStoreDto): Promise<Store> {
+    this.logger.log(`Creating a Store: ${storeDto}`);
 
-    const existingStore = await this.storeRepo.findOne({
-      where: [{ name: store.name }]
+    const store = await this.storeRepo.findOne({
+      where: { name: storeDto.name }
     });
 
-    if (existingStore) throw new ConflictException(`Esta Loja já existe!`);
+    if (store) {
+      throw new ConflictException(`A loja ${store} já existe!`);
+    }
 
-    const newStore = this.storeRepo.create(store);
+    const newStore = this.storeRepo.create(storeDto);
 
     return await this.storeRepo.save(newStore);
   }
@@ -49,20 +65,44 @@ export class StoreService implements IStoreService {
   async update(id: number, update: UpdateStoreDto): Promise<Store> {
     this.logger.log(`Get the Store of id ${id}.`);
 
-    await this.getOneById(id)
+    const store = await this.getOneById(id)
+    if (!store) {
+      throw new NotFoundException(`Loja ${id} não encontrada!`);
+    }
     await this.storeRepo.update({ id }, { ...update });
 
-    return await this.storeRepo.findOneBy({ id });
+    return await this.getOneById(id);
   }
 
   //Delete a store
   async delete(id: number): Promise<Store> {
     this.logger.log(`Deleting Store ${id}.`);
 
-    const store = this.getOneById(id);
-
+    const store = await this.getOneById(id);
+    if (!store) {
+      throw new NotFoundException(`Loja ${id} não encontrada!`);
+    }
+ 
     await this.storeRepo.delete({ id });
 
     return store;
+  }
+
+  //Verify the store owner
+  private async verifyOwner(owner: PartialUserDto): Promise<User | null> {
+    const user = await this.userService.getOne({...owner});
+    if (!user) {
+      return null;
+    }
+    return user;
+  }
+
+  //Verify the store category
+  private async verifyCategory(categoryId: number): Promise<Category | null> {
+    const category = await this.categoryService.getOneById(categoryId);
+    if (!category) {
+      return null;
+    }
+    return category;
   }
 }
